Quite large rewrite
Instead of setting drives and setting partition numbers on this drive
for the boot and UEFI partitions, enumerate boot and UEFI partitions
explicitly.
This way we can use partition uuids to pinpoint the exact partition on
a specific drive.
No chance of overwriting wrong partition.

When having multiple partitions; instead of setting up two mirrors for
the boot and UEFI partitions and then running nixos-rebuild(this does
not work with grub, unfortunately), run nixos-rebuild on the first
boot and UEFI partitions and, after this, then set up 2 mirrors with
the corresponding partitions and let these mirrors settle.

Leave in some debug messages, as this is quite untested...